feat(db): reconcile's currency pass through CrateStack, and why the provider pass cannot follow
`ConfigReconcile::reconcile` runs its currency read as
`find_unique(code).for_update().run_in_tx(tx, ctx)` and its currency write as
`upsert(CreateCurrencyInput).run_in_tx(tx, ctx)` — inside the transaction
`reconcile` already opened, after the same `pg_advisory_xact_lock`, in the
same sorted order. First use of `run_in_tx` anywhere in vpay: CrateStack joins
vpay's transaction rather than opening one of its own.

The read is the guard, not a prefetch. The statement it replaced was a
deliberate no-op `ON CONFLICT DO UPDATE SET exponent = currencies.exponent`
whose RETURNING handed back the *stored* value so Rust could compare it.
CrateStack renders `SET exponent = EXCLUDED.exponent` — the overwrite
`CurrencyExponentConflict` exists to refuse — so the comparison has to happen
before the write, under a row lock.

`model Currency` gains read/create/update `@@allow(auth().isSystem())`;
`model Provider` gains read only.

The provider pass stays hand-written, and that is measured rather than
pending: `cratestack-macros` drops every `@default(...)` field from both
`Create{Model}Input` and `upsert_update_columns`, and `model Provider` carries
one on all five capability booleans because migration 0002's table does. A
CrateStack upsert would write the column defaults instead of the deployment's
configuration and would never carry a capability change to an existing row —
a rail an operator had just disabled would come back enabled. Unblocking it
means dropping five DB defaults, which is a maintainer's decision and is
written up rather than taken.

`CurrencySeed::exponent` and `CurrencyExponentConflict`'s `stored`/`seeded`
are `i64` now (the column is BIGINT; sqlx refuses the narrowing rather than
performing it). `boot_seeds`' "exponent does not fit the column" arm became
unreachable by type, so the `try_from` is an `i64::from` and the `# Errors`
section says so.

Tests: four in `config_reconcile` (two `preview_sql` pins, the policy-slot
assertion, and a container-backed provider read/sqlx parity case that is what
the enum conversion is proven by), plus two container cases in
`tests/repositories.rs` — a hand-seeded exponent that must be read back and
refused, and a row-lock case that is red if `.for_update()` goes. That second
one exists because deleting `.for_update()` left all 103 `vpay-db` tests green.

+++ backends/crates/vpay-db/src/error.rs
@@ -166,8 +166,11 @@ pub enum DbError {
     /// decimal places while the database recorded zero.
     ///
     /// Refused rather than upserted, and this is the whole reason
-    /// `config_reconcile` reads the stored exponent back instead of writing
-    /// `DO UPDATE SET exponent = EXCLUDED.exponent`. The exponent is not a
+    /// `config_reconcile` reads the stored exponent — with
+    /// `find_unique(code).for_update()` since 2026-09-06 — before it upserts,
+    /// instead of letting `DO UPDATE SET exponent = EXCLUDED.exponent` land.
+    /// That statement is exactly what CrateStack's `upsert` renders, which is
+    /// why the read is not an optimisation there but the guard itself. The exponent is not a
     /// per-deployment setting; it is a property of the currency itself
     /// (`docs/flows/money.md`, and migration 0001's own comment), and every
     /// amount already stored is an integer count of minor units *at the
@@ -187,9 +190,19 @@ pub enum DbError {
         /// The ISO-4217 code whose exponent disagrees.
         code: String,
         /// What the database already holds, and what existing amounts mean.
-        stored: i32,
+        ///
+        /// `i64`, not `i32`, since migration 0032 widened
+        /// `currencies.exponent` to `BIGINT` — `Int` in a `.cstack` model
+        /// always emits `int8`, and an `int4` column is not a narrower `Int`
+        /// to `cratestack`, it is a column its introspector refuses to map at
+        /// all. The value is still bounded to `0..=4` two layers upstream
+        /// (`vpay_config`'s `Config::validate_all`, then
+        /// `currencies_exponent_range_check`); the width is about which
+        /// integer the column *is*, not about how large an exponent vpay
+        /// accepts.
+        stored: i64,
         /// What this deployment's configuration asked for.
-        seeded: i32,
+        seeded: i64,
     },
 
     /// A query that ran through CrateStack's data layer rather than through
